use getViewOnly
