--- packages/twenty-front/src/modules/object-record/record-board/states/isRecordBoardCompactModeActiveComponentState.ts
@@ -1,9 +1,0 @@
-import { RecordBoardComponentInstanceContext } from '@/object-record/record-board/states/contexts/RecordBoardComponentInstanceContext';
-import { createComponentState } from '@/ui/utilities/state/component-state/utils/createComponentState';
-
-export const isRecordBoardCompactModeActiveComponentState =
-  createComponentState<boolean>({
-    key: 'isRecordBoardCompactModeActiveComponentState',
-    defaultValue: false,
-    componentInstanceContext: RecordBoardComponentInstanceContext,
-  });
