+++ packages/twenty-front/src/modules/object-record/record-calendar/record-calendar-card/components/RecordCalendarCardHeader.tsx
@@ -1,11 +1,9 @@
 import { RecordChip } from '@/object-record/components/RecordChip';
 import { StopPropagationContainer } from '@/object-record/record-board/record-board-card/components/StopPropagationContainer';
-import { isRecordBoardCompactModeActiveComponentState } from '@/object-record/record-board/states/isRecordBoardCompactModeActiveComponentState';
 import { useRecordCalendarContextOrThrow } from '@/object-record/record-calendar/contexts/RecordCalendarContext';
 import { RecordCardHeaderContainer } from '@/object-record/record-card/components/RecordCardHeaderContainer';
-import { useRecordIndexIdFromCurrentContextStore } from '@/object-record/record-index/hooks/useRecordIndexIdFromCurrentContextStore';
 import { recordStoreFamilyState } from '@/object-record/record-store/states/recordStoreFamilyState';
-import { useRecoilComponentState } from '@/ui/utilities/state/component-state/hooks/useRecoilComponentState';
+import { useGetCurrentViewOnly } from '@/views/hooks/useGetCurrentViewOnly';
 
 import styled from '@emotion/styled';
 import { useRecoilValue } from 'recoil';
@@ -33,12 +31,9 @@ export const RecordCalendarCardHeader = ({
   const { objectMetadataItem } = useRecordCalendarContextOrThrow();
   const record = useRecoilValue(recordStoreFamilyState(recordId));
 
-  const { recordIndexId } = useRecordIndexIdFromCurrentContextStore();
+  const { currentView } = useGetCurrentViewOnly();
 
-  const [isCompactModeActive] = useRecoilComponentState(
-    isRecordBoardCompactModeActiveComponentState,
-    recordIndexId,
-  );
+  const isCompactModeActive = currentView?.isCompact ?? false;
 
   if (!isDefined(record)) {
     return null;
