+++ packages/twenty-front/src/modules/object-record/record-index/components/RecordIndexBoardDataLoaderEffect.tsx
@@ -1,10 +1,7 @@
 import { useEffect } from 'react';
-import { useRecoilValue } from 'recoil';
 
 import { contextStoreTargetedRecordsRuleComponentState } from '@/context-store/states/contextStoreTargetedRecordsRuleComponentState';
-import { isRecordBoardCompactModeActiveComponentState } from '@/object-record/record-board/states/isRecordBoardCompactModeActiveComponentState';
 import { recordBoardSelectedRecordIdsComponentSelector } from '@/object-record/record-board/states/selectors/recordBoardSelectedRecordIdsComponentSelector';
-import { recordIndexIsCompactModeActiveState } from '@/object-record/record-index/states/recordIndexIsCompactModeActiveState';
 import { useRecoilComponentValue } from '@/ui/utilities/state/component-state/hooks/useRecoilComponentValue';
 import { useSetRecoilComponentState } from '@/ui/utilities/state/component-state/hooks/useSetRecoilComponentState';
 
@@ -15,23 +12,10 @@ type RecordIndexBoardDataLoaderEffectProps = {
 export const RecordIndexBoardDataLoaderEffect = ({
   recordBoardId,
 }: RecordIndexBoardDataLoaderEffectProps) => {
-  const recordIndexIsCompactModeActive = useRecoilValue(
-    recordIndexIsCompactModeActiveState,
-  );
-
   const selectedRecordIds = useRecoilComponentValue(
     recordBoardSelectedRecordIdsComponentSelector,
     recordBoardId,
   );
-
-  const setIsCompactModeActive = useSetRecoilComponentState(
-    isRecordBoardCompactModeActiveComponentState,
-    recordBoardId,
-  );
-
-  useEffect(() => {
-    setIsCompactModeActive(recordIndexIsCompactModeActive);
-  }, [recordIndexIsCompactModeActive, setIsCompactModeActive]);
 
   const setContextStoreTargetedRecords = useSetRecoilComponentState(
     contextStoreTargetedRecordsRuleComponentState,
