--- packages/twenty-front/src/modules/object-record/record-index/states/recordIndexIsCompactModeActiveState.ts
@@ -1,5 +1,0 @@
-import { createState } from 'twenty-ui/utilities';
-export const recordIndexIsCompactModeActiveState = createState<boolean>({
-  key: 'recordIndexIsCompactModeActiveState',
-  defaultValue: false,
-});
